+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Fixed
-- Throw all types of errors. (Not only `AssertionError`)
+- Throw all types of errors which from Rize. (Not only `AssertionError`)
 
 ## [0.6.0] - 2018-03-24
 ### Added